Support |all| option when shimming DNS.lookup
Before, clients attempting an "all" style address lookup,
such as the SQL Server database module tedious does in its connector.js
when provided a hostname,
were liable to error out due to treating the argument as an array
when Replay was supplying a string. This might have looked like:

    TypeError: this.addresses.shift is not a function

Now, when the |all| option is detected, the string is wrapped in an
array to match client expectations.